+++ packages/shared/src/catalog/index.ts
@@ -5,6 +5,7 @@ export * from "./primitives/button";
 export * from "./primitives/link";
 export * from "./primitives/separator";
 export * from "./primitives/toggleButton";
+export * from "./primitives/toggleButtonGroup";
 export * from "./registry";
 export * from "./library";
 export * from "./componentCatalog";